backend/pyproject.toml
[project]
name = "jarvis-backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic[email]>=2.10.0",
    "pydantic-settings>=2.7.0",
    "PyJWT[crypto]>=2.10.0",
    "cryptography>=44.0.0",
    "slowapi>=0.1.9",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.12",
    "langgraph>=1.0.5",
    "langchain-core>=1.2.7",
    "langchain-deepseek>=1.0.1",
    "langchain-openai>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "qdrant-client>=1.12.0",
    "redis[hiredis]>=5.2.0",
    "minio>=7.2.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "tiktoken>=0.8.0",
    "httpx>=0.28.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
    "ruff>=0.8.0",
    "pyright>=1.1.390",
    "aiosqlite>=0.20.0",
    "psycopg2-binary>=2.9.11",
]

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "N",    # pep8-naming
    "C90",  # mccabe complexity
]
ignore = [
    "B008",  # FastAPI Depends() 在默认参数中调用函数，合法
    "N818",  # 允许自定义异常不以 Error 结尾
]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.pyright]
typeCheckingMode = "basic"
pythonVersion = "3.13"
exclude = ["alembic/", "tests/"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
